IN226106002_Gen-Ai/genai-project/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langgraph>=0.1.0
chromadb>=0.5.0
pypdf>=4.2.0
python-dotenv>=1.0.1
reportlab>=4.2.0
sentence-transformers>=3.0.0
streamlit>=1.30.0
